doc-retrieval/backend/requirements.txt
numpy
fastapi
uvicorn
pydantic
pymongo
faiss-cpu
sentence-transformers
